Refactor balance store with SQL queries in import route handler
